feat(bulk): little changes
refs: #1296
