+++ docs/medical-api/api-reference/document/bulk-document-download.mdx
@@ -6,7 +6,7 @@ api: "POST /medical/v1/document/bulk-download-url"
 
 When executed, this endpoint triggers an asynchronous process to generate download URLs for all the patient's documents. 
 
-The status of the bulk signing process is returned in the response. 
+The status of the bulk download process is returned in the response. 
 
 <Tip>
   To get updates about the bulk download process, check out our [webhooks guide](/medical-api/more-info/webhooks). Note that webhooks right now will return all available documents for a given patient.
