+++ packages/core/src/external/aws/lambda-logic/document-bulk-signing.ts
@@ -2,6 +2,9 @@ import { S3Utils } from "../s3";
 import { z } from "zod";
 import { DocumentReference } from "@medplum/fhirtypes";
 import { searchDocuments } from "../../opensearch/search-documents";
+import axios from "axios";
+import { capture } from "../../../util/notifications";
+const ossApi = axios.create();
 
 const SIGNED_URL_DURATION_SECONDS = 3000; // longer since a lot of docs
 
@@ -23,9 +26,11 @@ export type DocumentBulkSignerLambdaResponse = {
 export async function getSignedUrls(
   cxId: string,
   patientId: string,
+  requestId: string,
   bucketName: string,
-  region: string
-): Promise<DocumentBulkSignerLambdaResponse[]> {
+  region: string,
+  apiURL: string
+) {
   const s3Utils = new S3Utils(region);
 
   const documents: DocumentReference[] = await searchDocuments({ cxId, patientId });
@@ -62,7 +67,15 @@ export async function getSignedUrls(
     })
   );
 
-  return urls.filter(url => url !== undefined) as DocumentBulkSignerLambdaResponse[];
+  const response = urls.filter(url => url !== undefined) as DocumentBulkSignerLambdaResponse[];
+  const ossApiClient = apiClientBulkDownloadWebhook(apiURL);
+  ossApiClient.triggerWebhook({
+    cxId: cxId,
+    patientId: patientId,
+    requestId: requestId,
+    dtos: response,
+  });
+  console.log("internal called");
 }
 
 export const DocumentBulkSignerLambdaResponseSchema = z.object({
@@ -78,3 +91,36 @@ export const DocumentBulkSignerLambdaResponseSchema = z.object({
 export const DocumentBulkSignerLambdaResponseArraySchema = z.array(
   DocumentBulkSignerLambdaResponseSchema
 );
+
+export type BulkDownloadWebhookParams = {
+  cxId: string;
+  patientId: string;
+  requestId: string;
+  dtos: DocumentBulkSignerLambdaResponse[];
+};
+
+export function apiClientBulkDownloadWebhook(apiURL: string) {
+  const sendBulkDownloadUrl = `${apiURL}/internal/docs/triggerBulkDownloadWebhook`;
+
+  return {
+    triggerWebhook: async function (params: BulkDownloadWebhookParams) {
+      try {
+        console.log(`Trigger API on ${sendBulkDownloadUrl} w/ ${JSON.stringify(params)}`);
+        await ossApi.post(sendBulkDownloadUrl, params.dtos, {
+          params: { cxId: params.cxId, patientId: params.patientId, requestId: params.requestId },
+        });
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      } catch (error: any) {
+        const msg = "Error notifying API";
+        const extra = {
+          url: sendBulkDownloadUrl,
+          statusCode: error.response?.status,
+          error,
+        };
+        console.log(msg, extra);
+        capture.message(msg, { extra, level: "info" });
+        throw new Error(`Error from API: ${error.message}`);
+      }
+    },
+  };
+}
